Completed Directory Helper Included Null Validator Implemented XML Helper

+++ HelperProject/Helpers/DirectoryHelper.cs
@@ -12,12 +12,13 @@ namespace HelperProject.Helpers
         /// Helps in creating a directory if it does not exist or else will not do anything
         /// </summary>
         /// <param name="path">The full path of the directory</param>
-        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>
-        [CheckForNullParameters]
-        public ErrorHandler CreatDirectoryIfNotExists(string path)
+        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>        
+        public static ErrorHandler CreatDirectoryIfNotExists(string path)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(path))
+                    return new ErrorHandler("Arguments cannot be null");
                 string message;
                 if (!Directory.Exists(path))
                 {
@@ -41,12 +42,13 @@ namespace HelperProject.Helpers
         /// </summary>
         /// <param name="sourcePath">The full path of the directory where it currently exists</param>
         /// <param name="destinationPath">The full path of the directory where it has to be moved</param>
-        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>
-        [CheckForNullParameters]
-        public ErrorHandler MoveDirectoryIfExists(string sourcePath, string destinationPath)
+        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>        
+        public static ErrorHandler MoveDirectoryIfExists(string sourcePath, string destinationPath)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(sourcePath, destinationPath))
+                    return new ErrorHandler("Arguments cannot be null");
                 if (Directory.Exists(sourcePath))
                 {
                     Directory.Move(sourcePath, destinationPath);
@@ -68,12 +70,13 @@ namespace HelperProject.Helpers
         /// <param name="destinationPath">The full path of the directory where the file has to be moved</param>
         /// <param name="isForcedMove">Used to know if the file has to be moved even if the destination path does not exist.
         /// If destination path does not exist it is created.</param>
-        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>
-        [CheckForNullParameters]
-        public ErrorHandler MoveFileIfExists(string fileName, string sourcePath, string destinationPath, bool isForcedMove)
+        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>        
+        public static ErrorHandler MoveFileIfExists(string fileName, string sourcePath, string destinationPath, bool isForcedMove)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(fileName, sourcePath, destinationPath))
+                    return new ErrorHandler("Arguments cannot be null");
                 if (Directory.Exists(sourcePath))
                 {
                     var oldCombinedPath = Path.Combine(sourcePath, fileName);
@@ -124,12 +127,13 @@ namespace HelperProject.Helpers
         /// All the subdirectories and files in the directory will also be deleted.
         /// </summary>
         /// <param name="path">The full path of the directory</param>
-        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>
-        [CheckForNullParameters]
-        public ErrorHandler DeleteDirectoryIfExists(string path)
+        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>        
+        public static ErrorHandler DeleteDirectoryIfExists(string path)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(path))
+                    return new ErrorHandler("Arguments cannot be null");
                 if (Directory.Exists(path))
                 {
                     Directory.Delete(path, true);
@@ -148,12 +152,13 @@ namespace HelperProject.Helpers
         /// If there are any subdirectory or file in the directory, the directory will not be deleted.
         /// </summary>
         /// <param name="path">The full path of the directory</param>
-        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>
-        [CheckForNullParameters]
-        public ErrorHandler DeleteDirectoryIfEmpty(string path)
+        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>        
+        public static ErrorHandler DeleteDirectoryIfEmpty(string path)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(path))
+                    return new ErrorHandler("Arguments cannot be null");
                 if (Directory.Exists(path))
                 {
                     if (Directory.EnumerateDirectories(path).Count() <= 0 || Directory.EnumerateFiles(path).Count() <= 0)
@@ -177,12 +182,13 @@ namespace HelperProject.Helpers
         /// </summary>
         /// <param name="path">The full path of the directory where the file exists without the file name</param>
         /// <param name="fileName">The file name with the extension</param>
-        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>
-        [CheckForNullParameters]
-        public ErrorHandler DeleteFileIfExists(string path, string fileName)
+        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>        
+        public static ErrorHandler DeleteFileIfExists(string path, string fileName)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(path, fileName))
+                    return new ErrorHandler("Arguments cannot be null");
                 if (Directory.Exists(path))
                 {
                     var combinedPath = Path.Combine(path, fileName);
@@ -206,12 +212,13 @@ namespace HelperProject.Helpers
         /// </summary>
         /// <param name="path">The full path of the directory</param>
         /// <returns>Returns the list of all directory names at the specified path if any within a wrapper object.
-        /// Access it using Result property of the object returned</returns>
-        [CheckForNullParameters]
-        public ErrorHandler<List<string>> GetAllDirectoriesFromPath(string path)
+        /// Access it using Result property of the object returned</returns>        
+        public static ErrorHandler<List<string>> GetAllDirectoriesFromPath(string path)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(path))
+                    return new ErrorHandler<List<string>>("Arguments cannot be null");
                 if (Directory.Exists(path))
                 {
                     return new ErrorHandler<List<string>>(Directory.EnumerateDirectories(path).ToList(), "Successfully returned sub directory names in the directory");
@@ -232,12 +239,13 @@ namespace HelperProject.Helpers
         /// </summary>
         /// <param name="path">The full path of the directory</param>
         /// <returns>Returns the list of all file names at the specified path if any within a wrapper object.
-        /// Access it using Result property of the object returned</returns>
-        [CheckForNullParameters]
-        public ErrorHandler<List<string>> GetAllFilesFromPath(string path)
+        /// Access it using Result property of the object returned</returns>        
+        public static ErrorHandler<List<string>> GetAllFilesFromPath(string path)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(path))
+                    return new ErrorHandler<List<string>>("Arguments cannot be null");
                 if (Directory.Exists(path))
                 {
                     return new ErrorHandler<List<string>>(Directory.EnumerateFiles(path).ToList(), "Successfully returned file names in the directory");
@@ -261,12 +269,15 @@ namespace HelperProject.Helpers
         /// <param name="searchAsExact">To check if you want to search files with filenames which contains the search keyword or matches the exact keyword.
         /// If true, directory names that exactly matches the seaech keyword are returned or else the file names that contain the search keyword</param>
         /// <returns>Returns the list of matching directory names at the specified path if any within a wrapper object.
-        /// Access it using Result property of the object returned</returns>
-        [CheckForNullParameters]
-        public ErrorHandler<List<string>> SearchDirectoriesFromPath(string path, string searchText, bool searchAsExact)
+        /// Access it using Result property of the object returned</returns>        
+        public static ErrorHandler<List<string>> SearchDirectoriesFromPath(string path, string searchText = "", bool searchAsExact = false)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(path))
+                    return new ErrorHandler<List<string>>("Arguments cannot be null");
+                if (searchText == "")
+                    return GetAllDirectoriesFromPath(path);
                 if (Directory.Exists(path))
                 {
                     if (!searchAsExact)
@@ -293,12 +304,16 @@ namespace HelperProject.Helpers
         /// <param name="searchAsExact">To check if you want to search files with filenames which contains the search keyword or matches the exact keyword
         /// If true, file names that exactly matches the seaech keyword are returned or else the file names that contain the search keyword</param>
         /// <returns>Returns the list of matching file names at the specified path if any within a wrapper object.
-        /// Access it using Result property of the object returned</returns>
-        [CheckForNullParameters]
-        public ErrorHandler<List<string>> SearchFilesFromPath(string path, string searchText, bool searchAsExact)
+        /// Access it using Result property of the object returned.
+        /// Error Handler object which will also tell you if the operation succeeded or not. If not it gives the reason as a message</returns>        
+        public static ErrorHandler<List<string>> SearchFilesFromPath(string path, string searchText = "", bool searchAsExact = false)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(path))
+                    return new ErrorHandler<List<string>>("Arguments cannot be null");
+                if (searchText == "")
+                    return GetAllFilesFromPath(path);
                 if (Directory.Exists(path))
                 {
                     if (!searchAsExact)
@@ -317,11 +332,19 @@ namespace HelperProject.Helpers
             }
         }
 
-        [CheckForNullParameters]
-        public ErrorHandler RenameDirectoryIfExists(string path, string oldName, string newName)
+        /// <summary>
+        /// Helps to rename a directory is it exists
+        /// </summary>
+        /// <param name="path">The full path to the directory(excluding directory name)</param>
+        /// <param name="oldName">Old name of the directory</param>
+        /// <param name="newName">New name of the directory</param>
+        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>        
+        public static ErrorHandler RenameDirectoryIfExists(string path, string oldName, string newName)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(path, oldName, newName))
+                    return new ErrorHandler("Arguments cannot be null");
                 if (Directory.Exists(path))
                 {
                     var oldCombinedPath = Path.Combine(path, oldName);
@@ -345,11 +368,19 @@ namespace HelperProject.Helpers
             }
         }
 
-        [CheckForNullParameters]
-        public ErrorHandler RenameFileIfExists(string path, string oldFileName, string newFileName)
+        /// <summary>
+        /// Helps to rename a file is it exists
+        /// </summary>
+        /// <param name="path">The full path to the file(excluding file name)</param>
+        /// <param name="oldFileName">Old name of the file</param>
+        /// <param name="newFileName">New name of the file</param>
+        /// <returns>Error Handler object which will tell you if the operation succeeded or not. If not it gives the reason as a message</returns>        
+        public static ErrorHandler RenameFileIfExists(string path, string oldFileName, string newFileName)
         {
             try
             {
+                if (Validator.CheckParametersAreNull(path, oldFileName, newFileName))
+                    return new ErrorHandler("Arguments cannot be null");
                 if (Directory.Exists(path))
                 {
                     var oldCombinedPath = Path.Combine(path, oldFileName);
